requirements.txt
flask
requests
python-dotenv
gunicorn
opencv-python-headless
numpy